+++ src/main/java/io/github/demchaav/markdown/parser/FlexmarkMarkdownParser.java
@@ -1,5 +1,6 @@
 package io.github.demchaav.markdown.parser;
 
+import com.vladsch.flexmark.ext.emoji.EmojiExtension;
 import com.vladsch.flexmark.ext.footnotes.FootnoteExtension;
 import com.vladsch.flexmark.ext.gfm.strikethrough.StrikethroughExtension;
 import com.vladsch.flexmark.ext.gfm.tasklist.TaskListExtension;
@@ -27,7 +28,8 @@ public final class FlexmarkMarkdownParser {
         MutableDataSet options = new MutableDataSet();
         options.set(Parser.EXTENSIONS, List.of(
                 StrikethroughExtension.create(), TablesExtension.create(),
-                TaskListExtension.create(), FootnoteExtension.create()));
+                TaskListExtension.create(), FootnoteExtension.create(),
+                EmojiExtension.create()));
         this.parser = Parser.builder(options).build();
     }